# services/jurisdiction-rules-svc/openapi.yaml
openapi: "3.1.0"
info:
  title: Jurisdiction Rules Service
  version: "1.1.0"
  description: |
    Authoritative registry of jurisdiction identities and their effective-dated
    applicability rules (03-microservices.md §8.2).

    **Governance invariants:**
    - `rule_payload` holds APPLICABILITY METADATA ONLY — never computation values.
      Thresholds, rates and bands belong to the Tax and Payroll services (OQ-1, Model B).
    - Every type/status/domain field is a VARCHAR tag stored as data. New jurisdiction
      types, rule domains and status values are added by data migration, never by a
      code change (OQ-3).
    - Nothing is deleted. Jurisdictions are deactivated with `active_flag` + `effective_to`;
      rules are end-dated and moved to SUPERSEDED or RETIRED.
    - Historical actions must always be explainable against the rule set active at the
      time of execution, so point-in-time reads keep working after a jurisdiction is
      deactivated and after a rule is superseded.
    - `legal_drift_state` transitions are append-only. The column carries the current
      value; `/v1/rules/{id}/drift-events` carries the history (OQ-4).
    - Jurisdiction and rule data are platform-wide reference data. There is no
      `tenant_id` and no row-level security on these tables.

    **Read vs runtime resolution.** `/rules` is the raw per-jurisdiction, effective-dated
    view — the audit and replay surface, one jurisdiction only. `/rule-pack` is the
...

...

...
servers:
...
paths:
...
  /v1/rule-resolutions:resolve:
    post:
      tags: [Resolver]
      summary: Resolve which rule applies, from verified released pack artifacts
      description: |
        Action RULE_RESOLUTION_RESOLVE. Body: jurisdiction (a code: a FACT, not a conclusion), rule_domain,
        rule_code, effective_at (RFC3339), pack_ref? (restrict to one pack), pack_version? (with pack_ref:
        replay that exact version, for reproducing a past decision; WITHDRAWN/EMERGENCY_BLOCKED never resolve).
        Optional header Idempotency-Key (scoped to the caller; same key + same request returns the same
        decision with replayed=true; same key + different request is 409).

        The server picks the pack: among RELEASED versions (configurable; CERTIFIED only outside
        production) whose scope names the jurisdiction and whose window contains the instant, the highest
        version per pack. Every consulted pack is verified first (artifact digest, recorded digests,
        registered non-revoked key, signature, and the certification's own signature). If a covering pack
        fails verification the answer is **503 pack_unverified**, never a different answer from an older
        pack (JUR-NEG-03/04). Overlapping rules are decided only by declared precedence or supersedes;
        otherwise AMBIGUOUS. Two packs defining the same rule for the same jurisdiction and instant is a
        cross-pack conflict, also AMBIGUOUS (JUR-NEG-24). No eligible pack naming the jurisdiction is
        UNSUPPORTED_JURISDICTION (JUR-NEG-22): nothing is guessed.

        Every answer except 503 is recorded in the evidence ledger. HTTP: 200 RESOLVED or NO_RULE;
        409 AMBIGUOUS; 422 UNSUPPORTED_JURISDICTION; 503 pack_unverified. The body is
        {decision_id, replayed, decision{outcome, pack{pack_ref, version, artifact_digest, certification_id},
        rule{...payload}, sources[], basis, considered_rule_ids[], packs_consulted[], explanation[]}}.

        NOT BUILT: POST /jurisdiction-decisions:resolve (needs authoritative entity/counterparty/transaction
        facts and place-of-supply rules) and POST /regulatory-calculations:execute (needs a rule formula
        language). Neither is faked.
      operationId: resolveRule
      security: [{ principalHeader: [] }]
      parameters:
        - { $ref: '#/components/parameters/PrincipalID' }
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 128 } }
      responses:
        "200": { description: RESOLVED or NO_RULE. }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/MissingPrincipal' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { description: Pinned pack version not found. }
        "409": { description: 'AMBIGUOUS, idempotency_conflict, or pinned_pack_not_eligible.' }
        "422": { description: UNSUPPORTED_JURISDICTION. }
        "503": { description: pack_unverified (fail closed) or authorization/store unavailable. }
  /v1/rule-decisions/{decision_id}:
    get:
      tags: [Resolver]
      summary: A recorded decision (the evidence of a past resolution)
      description: "Action RULE_DECISION_VIEW. The stored response is exactly what the caller received."
      operationId: getRuleDecision
      security: [{ principalHeader: [] }]
      parameters: [{ name: decision_id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/PrincipalID' }]
      responses: { "200": { description: The decision. }, "404": { $ref: '#/components/responses/NotFound' } }
  /v1/pack-coverage:
    get:
      tags: [Resolver]
      summary: Which eligible pack versions name a jurisdiction (read model, no legal conclusion)
      operationId: packCoverage
      parameters:
        - { name: jurisdiction, in: query, required: true, schema: { type: string } }
        - { name: effective_at, in: query, schema: { type: string, format: date-time } }
      responses: { "200": { description: 'Packs naming the jurisdiction, with in_force.' }, "400": { $ref: '#/components/responses/BadRequest' } }
  /v1/admin/resolver-cache:invalidate:
    post:
      tags: [Admin, Resolver]
      summary: Drop every cached pack so the next resolution re-reads and re-verifies
      description: |
        Action RESOLVER_CACHE_INVALIDATE. Verified artifacts are cached for RESOLVER_CACHE_TTL_SECONDS
        (default 30, 0 = never). A key revocation or withdrawal takes effect within that bound; call this to
        make it immediate. The cache is per process: with several replicas, invalidate each.
      operationId: invalidateResolverCache
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }]
      responses: { "200": { description: Invalidated; cache counters in the body. }, "403": { $ref: '#/components/responses/Forbidden' } }

  # ── ZS-JUR-001 Wave 7: release, rings, rollback, hotfix, source-change intake, metrics ──
  # Every command below is its own authorization action. Refusals that depend on several gates answer
  # 409 {"error":"operation_blocked","reasons":[...]} listing ALL unmet gates at once.
...
components:
  securitySchemes:
    principalHeader:
      type: apiKey
      in: header
      name: X-Principal-Id
      description: |
        The acting principal, set by the gateway's ForwardAuth middleware after it
        verifies the caller's identity envelope against identity-svc's JWKS. Callers
        do not set this themselves — a request that reaches the service without it is
        rejected with `401 missing_principal` rather than attributed to the platform.
  parameters:
...
    PrincipalID:
      name: X-Principal-Id
      in: header
      required: true
      description: See the `principalHeader` security scheme.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: |
        Malformed or incomplete request. Codes include `missing_field`,
        `invalid_request_body`, `invalid_effective_at`, `invalid_effective_period`,
        `invalid_limit`, `invalid_offset`, `invalid_rule_status`,
        `invalid_rule_payload`, `invalid_status`, `invalid_drift_state`.
        An unknown JSON field is rejected rather than ignored.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    MissingPrincipal:
      description: '`missing_principal` — no caller identity on the request.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`unauthorized` — the Authorization Service denied the action.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        `jurisdiction_not_found` / `rule_not_found`. Also returned for a malformed id,
        which cannot name an existing row.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, instance]
      description: |
        RFC 9457 Problem Details format. Clients can switch on `type` or `title`.
        The `extension` fields provide machine-readable codes.
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Stable machine-readable code for client logic (e.g., "missing_field").
        field:
          type: string
          description: The offending field, when code is "missing_field".
        jurisdiction_id:
          type: string
          format: uuid
          description: The jurisdiction ID, when relevant.
...